update inline comments & merge the backoff
