+++ crates/stateless-common/src/lib.rs
@@ -2,7 +2,9 @@ pub mod logging;
 pub mod metrics;
 pub use metrics::{RpcMethod, RpcMetrics};
 pub mod rpc_client;
-pub use rpc_client::{RpcClient, RpcClientConfig, SetValidatedBlocksResponse, WitnessRequestKeys};
+pub use rpc_client::{
+    BackoffPolicy, RpcClient, RpcClientConfig, SetValidatedBlocksResponse, WitnessRequestKeys,
+};
 pub mod witness_size;
 pub use witness_size::{WitnessSizeBreakdown, estimate_witness_size};
